refactor: make node extensible, extract ipfs storage logic
