+++ packages/data-access/src/data-access.ts
@@ -37,6 +37,11 @@ export interface IDataAccessOptions {
    * Index of the ignored location with the reason
    */
   ignoredLocationIndex: IgnoredLocationIndex;
+
+  /**
+   * Specifies whether to start the synchronization on initialization
+   */
+  autoStartSynchronization: boolean;
 }
 
 const emptyChannelsWithTopics: DataAccessTypes.IReturnGetChannelsByTopic = {
@@ -66,6 +71,11 @@ export default class DataAccess implements DataAccessTypes.IDataAccess {
   // This object allows to handle the periodical call of the function
   private synchronizationTimer: IntervalTimer;
 
+  /**
+   * Specifies whether to start the synchronization on initialization
+   */
+  private autoStartSynchronization: boolean;
+
   // Timestamp of the last synchronization
   //
   // Are you debugging and this value is not changing as much as you think it should? Read bellow.
@@ -91,13 +101,21 @@ export default class DataAccess implements DataAccessTypes.IDataAccess {
       logger: new Utils.SimpleLogger(),
       synchronizationIntervalTime: DEFAULT_INTERVAL_TIME,
       transactionIndex: new TransactionIndex(),
+      autoStartSynchronization: false,
     };
-    const { ignoredLocationIndex, logger, synchronizationIntervalTime, transactionIndex } = {
+    const {
+      ignoredLocationIndex,
+      logger,
+      synchronizationIntervalTime,
+      transactionIndex,
+      autoStartSynchronization,
+    } = {
       ...defaultOptions,
       ...options,
     };
     this.storage = storage;
     this.lastSyncStorageTimestamp = 0;
+    this.autoStartSynchronization = autoStartSynchronization;
     this.synchronizationTimer = new IntervalTimer(
       () => this.synchronizeNewDataIds(),
       synchronizationIntervalTime,
@@ -144,6 +162,14 @@ export default class DataAccess implements DataAccessTypes.IDataAccess {
     await this.pushLocationsWithTopics(allDataWithMeta.entries);
 
     this.isInitialized = true;
+
+    if (this.autoStartSynchronization) {
+      this.startAutoSynchronization();
+    }
+  }
+
+  public async close(): Promise<void> {
+    this.stopAutoSynchronization();
   }
 
   /**
@@ -440,7 +466,9 @@ export default class DataAccess implements DataAccessTypes.IDataAccess {
    * Stop to synchronize with the storage automatically
    */
   public stopAutoSynchronization(): void {
-    this.synchronizationTimer.stop();
+    if (this.synchronizationTimer.isStarted) {
+      this.synchronizationTimer.stop();
+    }
   }
 
   /**
