+++ packages/ethereum-storage/src/ethereum-entries-to-ipfs-content.ts
@@ -5,7 +5,6 @@ import { getMaxIpfsReadRetry } from './config';
 
 import IgnoredDataIds from './ignored-dataIds';
 import IpfsConnectionError from './ipfs-connection-error';
-import IpfsManager from './ipfs-manager';
 
 // rate of the size of the Header of a ipfs file regarding its content size
 // used to estimate the size of a ipfs file from the content size
@@ -21,7 +20,7 @@ const SAFE_MAX_HEADER_SIZE = 500;
  */
 export default async function EthereumEntriesToIpfsContent(
   ethereumEntries: StorageTypes.IEthereumEntry[],
-  ipfsManager: IpfsManager,
+  ipfsStorage: StorageTypes.IIpfsStorage,
   ignoredDataIdsIndex: IgnoredDataIds,
   logger: LogTypes.ILogger,
   maxConcurrency: number,
@@ -60,7 +59,7 @@ export default async function EthereumEntriesToIpfsContent(
       // Reject on error when no file is found on IPFS
       // or when the declared size doesn't correspond to the size of the content stored on ipfs
       async (ethereumEntry: StorageTypes.IEthereumEntry) => {
-        return getIpfsContent(ethereumEntry, tryIndex + 1, ipfsManager, logger);
+        return getIpfsContent(ethereumEntry, tryIndex + 1, ipfsStorage, logger);
       },
       {
         concurrency: maxConcurrency,
@@ -77,7 +76,7 @@ export default async function EthereumEntriesToIpfsContent(
         // content found and not error
         finalIpfsContents.push(ipfsContent);
       } else if (entryWithError) {
-        const errorType = entryWithError.error!.type;
+        const errorType = entryWithError.error?.type;
         if (errorType === StorageTypes.ErrorEntries.INCORRECT_FILE) {
           incorrectFileCount++;
           // no retry needed, just store it
@@ -92,7 +91,7 @@ export default async function EthereumEntriesToIpfsContent(
           ethereumEntriesToProcess.push(entryWithError);
         } else {
           throw new Error(
-            `Unexpected Error for the hash: ${entryWithError.hash}, ${entryWithError.error?.type}, ${entryWithError.error?.message}`,
+            `Unexpected Error for the hash: ${entryWithError.hash}, ${errorType}, ${entryWithError.error?.message}`,
           );
         }
       }
@@ -139,7 +138,7 @@ export default async function EthereumEntriesToIpfsContent(
 async function getIpfsContent(
   ethereumEntry: StorageTypes.IEthereumEntry,
   tryIndex: number,
-  ipfsManager: IpfsManager,
+  ipfsStorage: StorageTypes.IIpfsStorage,
   logger: LogTypes.ILogger,
 ): Promise<{
   ipfsContent: StorageTypes.IEntry | null;
@@ -168,7 +167,7 @@ async function getIpfsContent(
   try {
     const startTime = Date.now();
     // Send ipfs request
-    ipfsObject = await ipfsManager.read(
+    ipfsObject = await ipfsStorage.read(
       ethereumEntry.hash,
       Number(ethereumEntry.feesParameters.contentSize) + ipfsHeaderMargin,
     );
