+++ packages/ethereum-storage/src/ethereum-storage.ts
@@ -2,12 +2,11 @@ import { LogTypes, StorageTypes } from '@requestnetwork/types';
 import Utils from '@requestnetwork/utils';
 import * as Bluebird from 'bluebird';
 import { EventEmitter } from 'events';
-import { getIpfsExpectedBootstrapNodes, getMaxConcurrency, getPinRequestConfig } from './config';
+import { getMaxConcurrency } from './config';
 
 import ethereumEntriesToIpfsContent from './ethereum-entries-to-ipfs-content';
 import EthereumMetadataCache from './ethereum-metadata-cache';
 import IgnoredDataIds from './ignored-dataIds';
-import IpfsManager from './ipfs-manager';
 import SmartContractManager from './smart-contract-manager';
 
 import * as Keyv from 'keyv';
@@ -27,10 +26,9 @@ export default class EthereumStorage implements StorageTypes.IStorage {
   public smartContractManager: SmartContractManager;
 
   /**
-   * Manager for IPFS
-   * This attribute is left public for mocking purpose to facilitate tests on the module
+   * Storage for IPFS
    */
-  public ipfsManager: IpfsManager;
+  private ipfsStorage: StorageTypes.IIpfsStorage;
 
   /**
    * Cache to store Ethereum metadata
@@ -71,7 +69,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
    */
   public constructor(
     externalBufferUrl: string,
-    ipfsGatewayConnection?: StorageTypes.IIpfsGatewayConnection,
+    ipfsStorage: StorageTypes.IIpfsStorage,
     web3Connection?: StorageTypes.IWeb3Connection,
     {
       getLastBlockNumberDelay,
@@ -90,7 +88,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
   ) {
     this.maxConcurrency = maxConcurrency || getMaxConcurrency();
     this.logger = logger || new Utils.SimpleLogger();
-    this.ipfsManager = new IpfsManager(ipfsGatewayConnection);
+    this.ipfsStorage = ipfsStorage;
     this.smartContractManager = new SmartContractManager(web3Connection, {
       getLastBlockNumberDelay,
       logger: this.logger,
@@ -132,23 +130,9 @@ export default class EthereumStorage implements StorageTypes.IStorage {
     }
 
     // Check IPFS node state - will throw in case of error
-    await this.checkIpfsNode();
+    await this.ipfsStorage.initialize();
 
     this.isInitialized = true;
-  }
-
-  /**
-   * Update gateway connection information and connect to the new gateway
-   * Missing value are filled with default config value
-   * @param ipfsConnection Information structure to connect to the ipfs gateway
-   */
-  public async updateIpfsGateway(
-    ipfsGatewayConnection: StorageTypes.IIpfsGatewayConnection,
-  ): Promise<void> {
-    this.ipfsManager = new IpfsManager(ipfsGatewayConnection);
-
-    // Check IPFS node state - will throw in case of error
-    await this.checkIpfsNode();
   }
 
   /**
@@ -177,32 +161,14 @@ export default class EthereumStorage implements StorageTypes.IStorage {
       throw new Error('Ethereum storage must be initialized');
     }
 
-    if (!content) {
-      throw Error('No content provided');
-    }
-
-    // Add content to IPFS and get the hash back
-    let ipfsHash: string;
-    try {
-      ipfsHash = await this.ipfsManager.add(content);
-    } catch (error) {
-      throw Error(`Ipfs add request error: ${error}`);
-    }
-
-    // Get content length from ipfs
-    let contentSize: number;
-    try {
-      contentSize = await this.ipfsManager.getContentLength(ipfsHash);
-    } catch (error) {
-      throw Error(`Ipfs get length request error: ${error}`);
-    }
+    const { ipfsHash, ipfsSize } = await this.ipfsStorage.ipfsAdd(content);
 
     const timestamp = Utils.getCurrentTimestampInSecond();
     const result: StorageTypes.IAppendResult = Object.assign(new EventEmitter(), {
       content,
       id: ipfsHash,
       meta: {
-        ipfs: { size: contentSize },
+        ipfs: { size: ipfsSize },
         local: { location: this.externalBufferUrl },
         state: StorageTypes.ContentState.PENDING,
         storageType: StorageTypes.StorageSystemType.LOCAL,
@@ -212,7 +178,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
     // store in the buffer the timestamp
     this.buffer[ipfsHash] = timestamp;
 
-    const feesParameters: StorageTypes.IFeesParameters = { contentSize };
+    const feesParameters: StorageTypes.IFeesParameters = { contentSize: ipfsSize };
 
     this.smartContractManager
       .addHashAndSizeToEthereum(ipfsHash, feesParameters)
@@ -222,7 +188,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
           id: ipfsHash,
           meta: {
             ethereum: ethereumMetadata,
-            ipfs: { size: contentSize },
+            ipfs: { size: ipfsSize },
             state: StorageTypes.ContentState.CONFIRMED,
             storageType: StorageTypes.StorageSystemType.ETHEREUM_IPFS,
             timestamp: ethereumMetadata.blockTimestamp,
@@ -241,51 +207,6 @@ export default class EthereumStorage implements StorageTypes.IStorage {
   }
 
   /**
-   * Add the content to ipfs
-   * To be used only in case of persisting the hash on ethereum outside the storage
-   *
-   * @param content Content to add into the storage
-   * @returns Promise resolving id used to retrieve the content
-   */
-  public async _ipfsAdd(data: string): Promise<StorageTypes.IIpfsMeta> {
-    if (!this.isInitialized) {
-      throw new Error('Ethereum storage must be initialized');
-    }
-
-    if (!data) {
-      throw Error('No data provided');
-    }
-
-    // Add a small check to at least having JSON data added
-    try {
-      JSON.parse(data);
-    } catch (error) {
-      throw Error(`data not JSON parsable: ${error}`);
-    }
-
-    // Add content to IPFS and get the hash back
-    let ipfsHash;
-    try {
-      ipfsHash = await this.ipfsManager.add(data);
-    } catch (error) {
-      throw Error(`Ipfs add request error: ${error}`);
-    }
-
-    // Get content length from ipfs
-    let ipfsSize;
-    try {
-      ipfsSize = await this.ipfsManager.getContentLength(ipfsHash);
-    } catch (error) {
-      throw new Error(`Ipfs get length request error: ${error}`);
-    }
-
-    return {
-      ipfsHash,
-      ipfsSize,
-    };
-  }
-
-  /**
    * Read content from the storage
    * @param Id Id used to retrieve content
    * @returns Promise resolving content from id
@@ -300,7 +221,6 @@ export default class EthereumStorage implements StorageTypes.IStorage {
 
     // Get Ethereum metadata
     let bufferTimestamp: number | undefined;
-    let ipfsObject;
 
     // Check if the data as been added on ethereum
     const ethereumMetadata = await this.ethereumMetadataCache.getDataIdMeta(id);
@@ -315,12 +235,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
       delete this.buffer[id];
     }
 
-    // Send ipfs request
-    try {
-      ipfsObject = await this.ipfsManager.read(id);
-    } catch (error) {
-      throw Error(`Ipfs read request error: ${error}`);
-    }
+    const ipfsObject = await this.ipfsStorage.read(id);
 
     const meta = ethereumMetadata
       ? {
@@ -380,9 +295,52 @@ export default class EthereumStorage implements StorageTypes.IStorage {
   public async getData(
     options?: StorageTypes.ITimestampBoundaries,
   ): Promise<StorageTypes.IEntriesWithLastTimestamp> {
-    const contentDataIdAndMeta = await this.getContentAndDataId(options);
+    if (!this.isInitialized) {
+      throw new Error('Ethereum storage must be initialized');
+    }
+    this.logger.info('Fetching dataIds from Ethereum', ['ethereum']);
+    const {
+      ethereumEntries,
+      lastTimestamp,
+    } = await this.smartContractManager.getEntriesFromEthereum(options);
 
-    return contentDataIdAndMeta;
+    // If no hash was found on ethereum, we return an empty list
+    if (!ethereumEntries.length) {
+      this.logger.info('No new data found.', ['ethereum']);
+      return {
+        entries: [],
+        lastTimestamp,
+      };
+    }
+
+    this.logger.debug('Fetching data from IPFS and checking correctness', ['ipfs']);
+
+    const entries = await ethereumEntriesToIpfsContent(
+      ethereumEntries,
+      this.ipfsStorage,
+      this.ignoredDataIds,
+      this.logger,
+      this.maxConcurrency,
+    );
+
+    const ids = entries.map((entry) => entry.id) || [];
+    // Pin data asynchronously
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    this.ipfsStorage.pinDataToIPFS(ids);
+
+    // Save existing ethereum metadata to the ethereum metadata cache
+    for (const entry of entries) {
+      const ethereumMetadata = entry.meta.ethereum;
+      if (ethereumMetadata) {
+        // PROT-504: The saving of dataId's metadata should be encapsulated when retrieving dataId inside smart contract (getPastEvents)
+        await this.ethereumMetadataCache.saveDataIdMeta(entry.id, ethereumMetadata);
+      }
+    }
+
+    return {
+      entries,
+      lastTimestamp,
+    };
   }
 
   /**
@@ -409,7 +367,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
 
     const entries = await ethereumEntriesToIpfsContent(
       ethereumEntries,
-      this.ipfsManager,
+      this.ipfsStorage,
       this.ignoredDataIds,
       this.logger,
       this.maxConcurrency,
@@ -417,8 +375,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
 
     const ids = entries.map((entry) => entry.id) || [];
     // Pin data asynchronously
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.pinDataToIPFS(ids);
+    void this.ipfsStorage.pinDataToIPFS(ids);
 
     // Save existing ethereum metadata to the ethereum metadata cache
     for (const entry of entries) {
@@ -433,35 +390,6 @@ export default class EthereumStorage implements StorageTypes.IStorage {
   }
 
   /**
-   * Pin an array of IPFS hashes
-   *
-   * @param hashes An array of IPFS hashes to pin
-   */
-  public async pinDataToIPFS(
-    hashes: string[],
-    {
-      delayBetweenCalls,
-      maxSize,
-      timeout,
-    }: StorageTypes.IPinRequestConfiguration = getPinRequestConfig(),
-  ): Promise<void> {
-    // How many slices we need from the total list of hashes to be under pinRequestMaxSize
-    const slices = Math.ceil(hashes.length / maxSize);
-
-    // Iterate over the hashes list, slicing it at pinRequestMaxSize sizes and pinning it
-    for (let i = 0; i < slices; i++) {
-      await new Promise<void>((res): NodeJS.Timeout => setTimeout(() => res(), delayBetweenCalls));
-      const slice = hashes.slice(i * maxSize, (i + 1) * maxSize);
-      try {
-        await this.ipfsManager.pin(slice, timeout);
-        this.logger.debug(`Pinned ${slice.length} hashes to IPFS node.`);
-      } catch (error) {
-        this.logger.warn(`Failed pinning some hashes the IPFS node: ${error}`, ['ipfs']);
-      }
-    }
-  }
-
-  /**
    * Get Information on the dataIds retrieved and ignored by the ethereum storage
    *
    * @param detailed if true get the list of the files hash
@@ -472,7 +400,7 @@ export default class EthereumStorage implements StorageTypes.IStorage {
     const dataIdsWithReason = await this.ignoredDataIds.getDataIdsWithReasons();
 
     const ethereum = this.smartContractManager.getConfig();
-    const ipfs = await this.ipfsManager.getConfig();
+    const ipfs = await this.ipfsStorage.getConfig();
 
     return {
       dataIds: {
@@ -486,96 +414,5 @@ export default class EthereumStorage implements StorageTypes.IStorage {
       },
       ipfs,
     };
-  }
-
-  /**
-   * Get all dataId and the contents stored on the storage
-   *
-   * @param options timestamp boundaries for the data id retrieval
-   * @returns Promise resolving object with content and dataId of stored data
-   */
-  private async getContentAndDataId(
-    options?: StorageTypes.ITimestampBoundaries,
-  ): Promise<StorageTypes.IEntriesWithLastTimestamp> {
-    if (!this.isInitialized) {
-      throw new Error('Ethereum storage must be initialized');
-    }
-    this.logger.info('Fetching dataIds from Ethereum', ['ethereum']);
-    const {
-      ethereumEntries,
-      lastTimestamp,
-    } = await this.smartContractManager.getEntriesFromEthereum(options);
-
-    // If no hash was found on ethereum, we return an empty list
-    if (!ethereumEntries.length) {
-      this.logger.info('No new data found.', ['ethereum']);
-      return {
-        entries: [],
-        lastTimestamp,
-      };
-    }
-
-    this.logger.debug('Fetching data from IPFS and checking correctness', ['ipfs']);
-
-    const entries = await ethereumEntriesToIpfsContent(
-      ethereumEntries,
-      this.ipfsManager,
-      this.ignoredDataIds,
-      this.logger,
-      this.maxConcurrency,
-    );
-
-    const ids = entries.map((entry) => entry.id) || [];
-    // Pin data asynchronously
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.pinDataToIPFS(ids);
-
-    // Save existing ethereum metadata to the ethereum metadata cache
-    for (const entry of entries) {
-      const ethereumMetadata = entry.meta.ethereum;
-      if (ethereumMetadata) {
-        // PROT-504: The saving of dataId's metadata should be encapsulated when retrieving dataId inside smart contract (getPastEvents)
-        await this.ethereumMetadataCache.saveDataIdMeta(entry.id, ethereumMetadata);
-      }
-    }
-
-    return {
-      entries,
-      lastTimestamp,
-    };
-  }
-
-  /**
-   * Verify the ipfs node (connectivity and network)
-   * Check if the node is reachable and if the list of bootstrap nodes is correct
-   *
-   * @returns nothing but throw if the ipfs node is not reachable or in the wrong network
-   */
-  private async checkIpfsNode(): Promise<void> {
-    // check ipfs connection - will throw in case of error
-    this.logger.info('Checking ipfs connection', ['ipfs', 'sanity']);
-    try {
-      await this.ipfsManager.getIpfsNodeId();
-    } catch (error) {
-      throw Error(`IPFS node is not accessible or corrupted: ${error}`);
-    }
-
-    // check if the ipfs node is in the request network private network - will throw in case of error
-    this.logger.info('Checking ipfs network', ['ipfs', 'sanity']);
-    try {
-      const bootstrapList = await this.ipfsManager.getBootstrapList();
-
-      const bootstrapNodeFoundCount: number = getIpfsExpectedBootstrapNodes().filter(
-        (nodeExpected) => bootstrapList.includes(nodeExpected),
-      ).length;
-
-      if (bootstrapNodeFoundCount !== getIpfsExpectedBootstrapNodes().length) {
-        throw Error(
-          `The list of bootstrap node in the ipfs config don't match the expected bootstrap nodes`,
-        );
-      }
-    } catch (error) {
-      throw Error(`IPFS node bootstrap node check failed: ${error}`);
-    }
   }
 }
