+++ packages/ethereum-storage/src/index.ts
@@ -1,1 +1,2 @@
 export { default as EthereumStorage } from './ethereum-storage';
+export { IpfsStorage } from './ipfs-storage';
