+++ packages/request-node/src/request/getTransactionsByChannelId.ts
@@ -1,12 +1,11 @@
-import { DataAccess } from '@requestnetwork/data-access';
-import { LogTypes } from '@requestnetwork/types';
+import { DataAccessTypes, LogTypes } from '@requestnetwork/types';
 import { Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
 
 const GET_TRANSACTIONS_TIMEOUT = 600000;
 
 export default class GetTransactionsByChannelIdHandler {
-  constructor(private logger: LogTypes.ILogger, private dataAccess: DataAccess) {
+  constructor(private logger: LogTypes.ILogger, private dataAccess: DataAccessTypes.IDataAccess) {
     this.handler = this.handler.bind(this);
   }
